requirements.txt
asgiref==3.8.0
Brotli==1.1.0
dj-database-url==2.1.0
Django==5.0.3
djangorestframework
gunicorn==21.2.0
packaging==24.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlparse==0.4.4
typing_extensions==4.10.0
whitenoise==6.6.0
django-cors-headers